backend/services/user-service/requirements.txt
# FastAPI Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Redis client
redis>=5.0.1

# Security & Authentication
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.9
email-validator>=2.1.0
